remove parse_appendvec_filename
